django-core/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.1
billiard==4.2.4
boto3==1.42.21
botocore==1.42.21
celery==5.6.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
django-cors-headers>=4.0,<5.0
contourpy==1.3.3
cron_descriptor==2.0.6
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
Django==5.2.9
django-celery-beat==2.8.1
django-environ==0.12.0
django-prometheus==2.4.1
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.127.0
fonttools==4.61.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kagglehub==0.4.0
kagglesdk==0.1.15
kiwisolver==1.4.9
kombu==5.6.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
gunicorn